Support for different syntaxes
Add a mechanism for defining different comment syntaxes.
Use this mechanism to define shell and C style comments
and add testcses for Python (shell style) and JS (C style).
Add a mechanism for controlling which files to add
the copyright statement to.
Update copyright statement to 2022 and remove the UTF-8
character to make it easier to get clean tests reghardless
of locale.